Refactor chained syntax to be transactional
Because cloning objects and arrays is generally O(N) in the size of the number
of keys or elements, immutable operations are relatively expensive. In cases
where multiple operations are performed without any opportunity for others to
observe intermediate values, the obvious optimization is to clone only once (for
the first operation) and mutate afterward.

So, refactor our "bound" or chained syntax as follows:
- given a chain of operations on a source object, clone any descendants of that
  source object at most once
- throw on any attempt to perform an operation after `value` is called on chain

Achieving this necessitated some refactoring of the internal API. All operations
now take a source (`src`) and a destination (`dest`). When called "unchained",
the mutation methods receive a `null` source.

Fix #5